+++ conda/platform_darwin_amd64.go
@@ -29,7 +29,7 @@ var (
 func CondaEnvironment() []string {
 	env := os.Environ()
 	env = append(env, fmt.Sprintf("MAMBA_ROOT_PREFIX=%s", common.RobocorpHome()))
-	tempFolder := RobocorpTemp()
+	tempFolder := common.RobocorpTemp()
 	env = append(env, fmt.Sprintf("TEMP=%s", tempFolder))
 	env = append(env, fmt.Sprintf("TMP=%s", tempFolder))
 	return env